src/lambda/requirements.txt
# All modules in src/lambda are part of the same Lambda package
liteLLM==1.71.3
python-jose 